+++ WindowsFormsApplication1/WindowsFormsApplication1.csproj
@@ -67,6 +67,7 @@
       <DependentUpon>RegistrationForm.cs</DependentUpon>
     </Compile>
     <Compile Include="TimeClass.cs" />
+    <Compile Include="Types.cs" />
     <EmbeddedResource Include="Authorization.resx">
       <DependentUpon>Authorization.cs</DependentUpon>
     </EmbeddedResource>
